% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatFWHM}
\alias{FeatFWHM}
\title{Find Full Width at Half maximum by spline interpolation}
\usage{
FeatFWHM(y, x = seq_along(y), n = 30 * length(x), method = "walk",
  basal = 1)
}
\arguments{
\item{y}{numeric, typically measured variable, data must be normalized!}

\item{x}{numeric, typically time vector}

\item{n}{numeric, number of points returned by interpolation. Increase improves accuracy.}

\item{method}{one of "minimum" or "walk". Walk method is highly recommended as it avoids many pitfalls
of the minimum method. Minimum simply pick the points which values are the closest to the half of the max value.
Walk method walks left and right from the maximum point, and stops whenever half max value is crossed.
Data must be normalized, such that "basal" represents the activity out of excitation.}
}
\value{
A list of 3: left and right represent the value of x at which half of the maximum in y is reached.
fwhm is the difference between left and right
}
\description{
Find Full Width at Half maximum by spline interpolation
}
