% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatHalfMaxGrow}
\alias{FeatHalfMaxGrow}
\title{FeatHalfMaxGrow}
\usage{
FeatHalfMaxGrow(y, ...)
}
\arguments{
\item{y}{numeric, typically measured variable, data must be normalized!}

\item{...}{extra arguments for FeatFWHM}
}
\value{
Slope of the linear regression
}
\description{
Fit a linear model on growing phase of a peak between (possibly interpolated) time at which half-max of the peak is reached
and maximum of the peak
}
