% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatLagGrow}
\alias{FeatLagGrow}
\title{FeatLagGrow}
\usage{
FeatLagGrow(y, start)
}
\arguments{
\item{y}{numeric, typically measured variable, data must be normalized!}

\item{start}{int, at which index should the regression start.}
}
\value{
Slope of the linear regression
}
\description{
Fit a linear model on the growing phase, starting at specified time, til max of the peak.
}
