% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatMaxAmplitude}
\alias{FeatMaxAmplitude}
\title{FeatMaxAmplitude}
\usage{
FeatMaxAmplitude(y, basal = 1)
}
\arguments{
\item{y}{a numerical vector}

\item{basal}{value to be subtracted for the signal to obtain height of the peak as amplitude instead of absolute value}
}
\value{
list of 2:
\itemize{
  \item max: maximum value of y after subtracting basal from it.
  \item time.max: index at which $max is reached.
}
}
\description{
Subtracts a constant value to a vector and return the maximum value of the difference.
}
