% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatTrend_extrema}
\alias{MPFeatTrend_extrema}
\title{MPFeatTrend_extrema}
\usage{
MPFeatTrend_extrema(x, window.size, what, robust = FALSE)
}
\arguments{
\item{x}{a numerical vector. Trajectory with multiple peaks.}

\item{window.size}{integer, width of window for maxima detection.
See ?detect.peak.}

\item{what}{character, one of "minima" or "maxima". Define whether regression
should be performed on maxima or minima of peaks.}

\item{robust}{logical. If TRUE perform robusts linear regression.
Median-Based Linear Models are used, see ?mblm::mblm}
}
\value{
A list of 2. $trend comprises the fit object
}
\description{
Isolate local extrema and perform a linear regression.
Aimed at at giving an estimation of long-term trend in multi-peak signals.
}
\examples{

}
